pyproject.toml
[tool.poetry]
name = "tech-store"
version = "0.1.0"
description = "Tech store"
authors = ["fj-fj-fj <x1x1xx0@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
django-configurations = "^2.2"
loguru = "^0.5.3"
psycopg2-binary = "^2.9.1"
dj-database-url = "^0.5.0"
django-admin-honeypot = "^1.1.0"

[tool.poetry.dev-dependencies]
coverage = "^5.5"
flake8 = "^3.9.2"
flake8-functions = "^0.0.6"
flake8-functions-names = "^0.0.5"
flake8-variables-names = "^0.0.4"
django-stubs = "^1.8.0"
ipython = "^7.26.0"
isort = "^5.9.3"
mypy = "^0.910"
pytest = "^6.2.4"
ptpython = "^3.0.19"
snoop = "^0.4.0"
birdseye = "^0.9.2"
django-debug-toolbar = "^3.2.2"
django-extensions = "^3.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
